+++ config/subscription.php
@@ -32,15 +32,6 @@ return [
         'BonusCreditAddedNotification'                => \Domain\Subscriptions\Notifications\BonusCreditAddedNotification::class,
     ],
 
-    /*
-     * Active drivers to handle subscription
-     */
-    'available_drivers' => [
-        'paystack',
-        'paypal',
-        'stripe',
-    ],
-
     'metered_billing'   => [
         'settlement_period' => 30,
     ],
@@ -55,8 +46,6 @@ return [
             '52.31.139.75',
         ],
     ],
-
-    'driver' => env('SUBSCRIPTION_DRIVER', 'stripe'),
 
     'is_demo' => env('APP_DEMO', false),
 ];