+++ config/vuefilemanager.php
@@ -1,7 +1,7 @@
 <?php
 
 return [
-    'version' => '2.0.11',
+    'version' => '2.0.12',
 
     'is_demo' => env('APP_DEMO', false),