+++ resources/js/components/Subscription/ChangeSubscriptionPopup.vue
@@ -141,6 +141,16 @@ export default {
 					.then((response) => {
 						window.location = response.data.url
 					})
+					.catch((error) => {
+						if (error.response.status === 500 && error.response.data.type) {
+							events.$emit('alert:open', {
+								title: error.response.data.title,
+								message: error.response.data.message,
+							})
+						} else {
+							this.$isSomethingWrong()
+						}
+					})
 			}
 
 			// Change active subscription